Deduplicate fraud event IDs and fix payout export/count filters
