+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/payouts/payout-table.tsx
@@ -101,7 +101,6 @@ export function PayoutTable() {
   const {
     id: workspaceId,
     slug: workspaceSlug,
-    plan,
     defaultProgramId,
   } = useWorkspace();
