+++ apps/web/lib/zod/schemas/payouts.ts
@@ -73,6 +73,7 @@ export const payoutsCountQuerySchema = payoutsQuerySchema
     partnerId: true,
     invoiceId: true,
     groupId: true,
+    tenantId: true,
   })
   .extend({
     programId: z.string().optional(),
